refactor: move Locale utility to a trait and apply it to PageResource
This commit refactors the `Locale` utility class into a `HasLocales` trait. The trait is then applied to the `PageResource` to provide access to the sorted locales.

This change improves code organization and promotes reusability of the locale sorting logic. It also simplifies the `PageResource` by removing the direct dependency on the `Locale` class.
